updat for infomaniak env

+++ updat.sh
@@ -1,7 +1,6 @@
 #!/bin/bash
 
-# EXAMPLE CONFIG (see updat.yml)
-
+home=$(echo $HOME)
 user=""
 type=""
 admin_username="admin"
@@ -10,12 +9,15 @@ repository=""
 domain=""
 php_ver=""
 web_dir="www"
+destination="www"
 backup_files=""
 post_install_cmd=""
 composer_version=""
 git_username=""
 git_token=""
+rsa_file="$home/.ssh/id_rsa"
 tstamp=$(date +"%s")
+
 min_disk_usage=5000000 # ~ 5GB, used inside check_server_disk_usage
 
 export COMPOSER_ALLOW_SUPERUSER=1
@@ -57,7 +59,7 @@ kill_self(){
 check_args() {
   for arg in $@; do
     if [[ $arg == "-h" ]]; then
-      echo "Usage: updat [-h] [-f] [-ni] [--no-yarn] [--no-composer] [--no-install]"
+      echo "Usage: updat [-h] [-f] [-ni] [--no-pnpm] [--no-composer] [--no-install]"
       echo "  updat               : run the updat script"
       echo "  updat init          : create a sample updat.yml file inside current directory"
       echo "  updat self-update   : update the updat script"
@@ -65,9 +67,9 @@ check_args() {
       echo "  -h                  : show this message"
       echo "  -f                  : force launch, remove .updatlock file if existing"
       echo "  -ni                 : no interactions, don't prompt disk usage and do not ask to validate (old directory must be erased manually)"
-      echo "  --no-yarn           : copy existing node_modules directory instead of running yarn install command"
+      echo "  --no-pnpm           : copy existing node_modules directory instead of running pnpm install command"
       echo "  --no-composer       : copy existing vendor directory instead of running composer install command"
-      echo "  --no-install        : alias for --no-yarn + --no-composer"
+      echo "  --no-install        : alias for --no-pnpm + --no-composer"
       echo "  --no-status         : wont perform and show git status"
       kill_self
     fi
@@ -77,14 +79,14 @@ check_args() {
     if [[ $arg == "-ni" ]]; then
       no_interaction=1
     fi
-    if [[ $arg == "--no-yarn" ]]; then
-      no_yarn=1
+    if [[ $arg == "--no-pnpm" ]]; then
+      no_pnpm=1
     fi
     if [[ $arg == "--no-composer" ]]; then
       no_composer=1
     fi
     if [[ $arg == "--no-install" ]]; then
-      no_yarn=1
+      no_pnpm=1
       no_composer=1
     fi
     if [[ $arg == "--no-status" ]]; then
@@ -129,26 +131,10 @@ check_variable() {
 
 get_php_version() {
   # Get PHP version from running php -v with user
-  php_version=$(runuser -l "$user" -c 'php -v | tr " " "\n" | head -n 2 | tail -n 1')
+  php_version=$(php -v | tr " " "\n" | head -n 2 | tail -n 1)
   version=$(echo "$php_version" | tr "." "\n" | head -n 1)
   major_version=$(echo "$php_version" | tr "." "\n" | head -n 2 | tail -n 1)
   echo "$version.$major_version"
-}
-
-get_composer_version(){
-  if [[ -z "$composer_version" ]]; then
-      # set composer version following contao version
-      if [[ "$type" == "contao" ]]; then
-        contao_version=$(cat "$temp_install_dir/composer.json" | grep manager-bundle | tr '"' '\n' | tail -n 2 | head -n 1 | sed "s#\^##g")
-        if [[ -z $(echo "$contao_version" | grep "4\.1") ]]; then
-          composer_version=1
-        fi
-      fi
-      if [[ -z "$composer_version" ]]; then
-        composer_version=2
-      fi
-    fi
-    echo "$composer_version"
 }
 
 check_server_disk_usage() {
@@ -157,17 +143,16 @@ check_server_disk_usage() {
   project_size=$(du -s "$install_dir" | tr "\t" "\n" | head -n 1)
   project_size_h=$(du -hs "$install_dir" | tr "\t" "\n" | head -n 1)
 
-  df_infos=$(df "/home/$user" | tail -n1 | tr " " "\n" | grep -Ev "^$")
-  df_h_infos=$(df -h "/home/$user" | tail -n1 | tr " " "\n" | grep -Ev "^$")
+  df_infos=$(quota 2>&1 | tail -n 1 | tr " " "\n" | grep -Ev "^$")
+  df_h_infos=$(quota -s 2>&1 | tail -n 1 | tr " " "\n" | grep -Ev "^$")
 
-  space_left=$(echo "$df_infos" | head -n 4 | tail -n 1)
-  space_left_h=$(echo "$df_h_infos" | head -n 4 | tail -n 1)
-  space_used=$(echo "$df_infos" | head -n 3 | tail -n 1)
-  space_used_h=$(echo "$df_h_infos" | head -n 3 | tail -n 1)
+  space_used=$(echo "$df_infos" | head -n 1 | tail -n 1)
+  space_used_h=$(echo "$df_h_infos" | head -n 1 | tail -n 1)
   space_total=$(echo "$df_infos" | head -n 2 | tail -n 1)
+  space_total_h=$(echo "$df_h_infos" | head -n 2 | tail -n 1)
 
-  spacestring="[ Disk space : \e[31m$space_used_h used\e[0m | \e[33m~$project_size_h project size\e[0m | \e[32m$space_left_h left\e[0m ]"
-  space_string_escaped="[ Disk space : $space_used_h used | ~$project_size_h project size | $space_left_h left ]"
+  spacestring="[ Disk space : \e[31m$space_used_h used\e[0m | \e[33m~$project_size_h project size\e[0m | \e[32m$space_total_h total\e[0m ]"
+  space_string_escaped="[ Disk space : $space_used_h used | ~$project_size_h project size | $space_total_h total ]"
 
   length=${#space_string_escaped}
   length_used=$(awk "BEGIN {print $space_used / $space_total * $length}" | sed -E "s/\..*$//g")
@@ -190,13 +175,6 @@ check_server_disk_usage() {
   echo ""
 
   echo -e "$spacestring"
-
-  if [[ $(($space_left - $project_size)) -gt $min_disk_usage ]]; then
-    echo "" >/dev/null
-  else
-    echo "Fatal error : not enough disk space to perform update !"
-    exit
-  fi
 }
 
 parse_yml() {
@@ -243,12 +221,12 @@ load_config() {
   check_variable "user"
   check_variable "type"
   check_variable "domain"
-  check_variable "web_dir"
+  check_variable "destination"
+  destination=$(echo $destination | sed "s#~#$home#g")
 
   if [[ -z "$php_ver" ]]; then
     php_ver=$(get_php_version "$user")
   fi
-  #echo -e "\e[32mphp_ver:\e[0m $php_ver"
 
   # parse repository branch if given
   if [[ -z $(echo "$repository" | sed -e "s/^[^:]*//g") ]]; then
@@ -261,76 +239,81 @@ load_config() {
   check_variable "repository"
   check_variable "repository_branch"
 
-  backup_folder="/home/$user/_backup"
-  install_dir="/home/$user/$web_dir"
-  temp_install_dir="/home/$user/$web_dir""_$tstamp"
-  temp_old_install_dir="/home/$user/old_$web_dir""_$tstamp"
-
+  install_dir=$(dirname "$destination")
+  backup_folder="$install_dir/_backup"
+  web_dir=$(basename "$destination")
+  temp_install_dir="$install_dir/$web_dir""_$tstamp"
+  temp_old_install_dir="$install_dir/old_$web_dir""_$tstamp"
 
   # Show git status
-  cd "$install_dir"
-  if [[ -z "$no_status" ]]; then
-    git fetch 2>&1 > /dev/null | hilite "Fetching $repository:$repository_branch"
-    git_status=$(git status -sb)
-    if [[ $(echo "$git_status" | grep "\[") ]]; then
-        echo -e "${YELLOW}Projet non à jour avec${NC} $repository:$repository_branch"
-        git status -sb
-      else
-        echo -e "${GREEN}Projet à jour sur ${NC} $repository:$repository_branch"
+  if [[ -d "$destination" ]]; then
+    cd "$destination"
+    if [[ -z "$no_status" ]]; then
+      git fetch 2>&1 > /dev/null | hilite "Fetching $repository:$repository_branch"
+      git_status=$(git status -sb)
+      if [[ $(echo "$git_status" | grep "\[") ]]; then
+          echo -e "${YELLOW}Projet non à jour avec${NC} $repository:$repository_branch"
+          git status -sb
+        else
+          echo -e "${GREEN}Projet à jour sur ${NC} $repository:$repository_branch"
+      fi
     fi
+  else
+    echo "Nouveau projet"
   fi
   cd "$current_directory"
-
 
   if [[ -z "$no_interaction" ]]; then
     disk_usage=$(check_server_disk_usage)
     echo "$disk_usage"
-    read -ep $'Updating \e[32m'$domain$'\e[0m, a \e[32mphp'$php_ver-$type$'\e[0m project \e[32m'$user'@'$install_dir$'\e[0m (Press <Enter> to continue)'
+    read -ep $'Updating \e[32m'$domain$'\e[0m, a \e[32mphp'$php_ver-$type$'\e[0m project \e[32m'$user'@'$destination$'\e[0m (Press <Enter> to continue)'
   fi
 }
 
 php_ver_composer() {
-  "/usr/local/share/php$php_ver/bin/php" -d memory_limit=-1 "/usr/local/bin/composer" $@ -n
+  composer $@ -n
+  # "/usr/local/share/php$php_ver/bin/php" -d memory_limit=-1 "/usr/local/bin/composer" $@ -n
 }
 
 backup_save() {
   # Copy files directory
-  path="$1"
-  source="$install_dir/$path"
-  destination="$backup_folder/$path"
-  if [ -d "$source" ] || [ -f "$source" ]; then
-    echo "Saving $source to $destination"
+  backup_path="$1"
+  backup_source="$destination/$backup_path"
+  backup_destination="$backup_folder/$backup_path"
+
+  if [ -d "$backup_source" ] || [ -f "$backup_source" ]; then
+    echo "Saving $backup_source to $backup_destination"
   fi
 
-  if [[ -d "$source" ]]; then
-    mkdir -p $(dirname "$destination")
-    destination=$(dirname "$destination")
-    cp -R "$source" "$destination"
+  if [[ -d "$backup_source" ]]; then
+    mkdir -p $(dirname "$backup_destination")
+    backup_destination=$(dirname "$backup_destination")
+    cp -R "$backup_source" "$backup_destination"
   fi
-  if [[ -f "$source" ]]; then
-    mkdir -p $(dirname "$destination")
-    cp "$source" "$destination"
+  if [[ -f "$backup_source" ]]; then
+    mkdir -p $(dirname "$backup_destination")
+    cp "$backup_source" "$backup_destination"
   fi
 }
 
 backup_load() {
-  path="$1"
-  source="$backup_folder/$path"
-  destination="$temp_install_dir/$path"
+  backup_path="$1"
+  backup_source="$backup_folder/$backup_path"
+  backup_destination="$temp_install_dir/$backup_path"
 
-  if [ -d "$source" ] || [ -f "$source" ]; then
-    echo "Loading $source to $destination"
+  if [ -d "$backup_source" ] || [ -f "$backup_source" ]; then
+    echo "Loading $backup_source to $backup_destination"
   fi
 
-  if [[ -d "$source" ]]; then
-    mkdir -p $(dirname "$destination")
-    destination=$(dirname "$destination")
-    cp -R "$source" "$destination"
+  if [[ -d "$backup_source" ]]; then
+    mkdir -p $(dirname "$backup_destination")
+    backup_destination=$(dirname "$backup_destination")
+    cp -R "$backup_source" "$backup_destination"
   fi
 
-  if [[ -f "$source" ]]; then
-    mkdir -p $(dirname "$destination")
-    cp "$source" "$destination"
+  if [[ -f "$backup_source" ]]; then
+    mkdir -p $(dirname "$backup_destination")
+    cp "$backup_source" "$backup_destination"
   fi
 
 }
@@ -341,7 +324,7 @@ save_local_files() {
   for files in $backup; do
     backup_save "$files"
   done
-  chown -R $user. "/home/$user"
+  chown -R $user. "$install_dir"
 }
 
 load_local_files() {
@@ -350,7 +333,7 @@ load_local_files() {
   done
 }
 
-# nicer output for composer/yarn installs
+# nicer output for composer/pnpm installs
 hilite() {
   error=""
   name="$1"
@@ -423,7 +406,7 @@ post_install() {
     eval "$post_install_cmd"
   fi
 
-  chown -R $user. "/home/$user"
+  chown -R $user. "$install_dir"
 }
 contao_last_log() {
   project_dir="$1"
@@ -445,7 +428,7 @@ contao_last_log() {
 }
 custom_log() {
   if [[ "$type" == "contao" ]]; then
-    contao_last_log "$install_dir"
+    contao_last_log "$destination"
   fi
 }
 
@@ -467,9 +450,9 @@ apache_last_error_log() {
 }
 deployment_test(){
   rm -rf "$temp_old_install_dir"
-  mv "$install_dir" "$temp_old_install_dir"
-  mv "$temp_install_dir" "$install_dir"
-  cd "$install_dir"
+  mv "$destination" "$temp_old_install_dir"
+  mv "$temp_install_dir" "$destination"
+  cd "$destination"
 
   echo "Await response from $domain..."
   response=$(curl -L -k --write-out '%{http_code}' --silent --output /dev/null "$domain")
@@ -480,17 +463,19 @@ deployment_test(){
   if [[ "$response" -ne "200" ]]; then
     echo "Fatal error : [$response] response while loading updated site, reverting..."
 
-    custom_log
-    apache_last_error_log
+#    custom_log
+#    apache_last_error_log
 
-    mv "$install_dir" "$temp_install_dir"
-    mv "$temp_old_install_dir" "$install_dir"
+    mv "$destination" "$temp_install_dir"
+    mv "$temp_old_install_dir" "$destination"
   else
     if [[ "$no_interaction" ]]; then
       echo "" >/dev/null
     else
-      read -p "Do you want to remove previous version website files ? (Press <Enter> to continue, <CTRL+C> to cancel)"
-      rm -rf "$temp_old_install_dir"
+      if [[ -d "$temp_old_install_dir" ]]; then
+        read -p "Do you want to remove previous version website files ? (Press <Enter> to continue, <CTRL+C> to cancel)"
+        rm -rf "$temp_old_install_dir"
+      fi
     fi
   fi
 }
@@ -503,7 +488,7 @@ main(){
 
   # Init Bitbucket SSH Key
   eval $(ssh-agent -s -t 120) > /dev/null 2>&1
-  ssh-add /root/.ssh/bitbucket_rsa 2>&1 | hilite "Initializing SSH Agent" "ssh"
+  ssh-add "$rsa_file" 2>&1 | hilite "Initializing SSH Agent" "ssh"
 
   # Load config duh
   load_config 2>&1
@@ -516,7 +501,7 @@ main(){
   # Clone project
   rm -rf "$temp_install_dir"
   if [[ -z "$git_token" ]]; then
-    git clone "https://bitbucket.org/$repository" -b "$repository_branch" "$temp_install_dir" -vvv 2>&1 | hilite "Git clone $repository" "git"
+    git clone "git@bitbucket.org:$repository" -b "$repository_branch" "$temp_install_dir" -vvv 2>&1 | hilite "Git clone $repository" "git"
   else
     git clone "https://$git_username:$git_token@bitbucket.org/$repository" -b "$repository_branch" "$temp_install_dir" -vvv 2>&1 | hilite "Git clone $repository" "git"
   fi
@@ -527,11 +512,9 @@ main(){
   # Composer build
   cd "$temp_install_dir"
   if [[ -f "composer.json" ]]; then
-    composer_version=$(get_composer_version)
-    composer self-update --"$composer_version" -q -n
 
     if [[ -n "$no_composer" ]]; then
-      cp -r "$install_dir/vendor" "$temp_install_dir/vendor" | hilite "Composer copy" "composer"
+      cp -r "$destination/vendor" "$temp_install_dir/vendor" | hilite "Composer copy" "composer"
     else
       # install using user's php version
       php_ver_composer install --no-dev 2>&1 | hilite "Composer install using php$php_ver" "composer"
@@ -540,13 +523,13 @@ main(){
 
   # Npm build
   if [[ -f "package.json" ]]; then
-    if [[ -n "$no_yarn" ]]; then
-      cp -r "$install_dir/node_modules" "$temp_install_dir/node_modules" | hilite "YARN copy" "yarn"
+    if [[ -n "$no_pnpm" ]]; then
+      cp -r "$destination/node_modules" "$temp_install_dir/node_modules" | hilite "PNPM copy" "pnpm"
     else
-      yarn 2>&1 | hilite "YARN install" "yarn"
+      pnpm install 2>&1 | hilite "PNPM install" "pnpm"
     fi
 
-    yarn run build 2>&1 | hilite "YARN build" "yarn"
+    pnpm run build 2>&1 | hilite "PNPM build" "pnpm"
   fi
 
   # post install scripts
